UNTESTED-new tag for images being tested so will be built
docker-compose up will build new images only if both are true:
- there is a build: directive in the yml file
- an existing built image with the given tag doesn't exist

so need to change docker-compose-staging with a special tag for
a 'built' version that exists only on staging, to force docker
to build from the new code.

haven't tested if this works the way i think it does yet.